+++ src/deploy.js
@@ -84,6 +84,17 @@ module.exports = function deploy() {
           choices: ['auto', 'ignore', 'always'],
           default: 'auto',
         })
+        .option('minify', {
+          describe: 'Enables minification of the final action bundle.',
+          type: 'boolean',
+          default: false,
+        })
+        .option('svc-resolve-git-ref', {
+          alias: 'svcResolveGitRef',
+          describe: 'Service name for git-resolve-ref service',
+          type: 'string',
+          default: 'helix-services/resolve-git-ref@v1',
+        })
         .array('default')
         .nargs('default', 2)
         .coerce('default', arg => arg.reduce((result, value, index, array) => {
@@ -95,7 +106,7 @@ module.exports = function deploy() {
         }, {}))
         .group(['auto', 'wsk-auth', 'wsk-namespace', 'default', 'dirty'], 'Deployment Options')
         .group(['wsk-host', 'loggly-host', 'loggly-auth', 'target'], 'Advanced Options')
-        .group(['package', 'target'], 'Package options')
+        .group(['package', 'minify', 'target'], 'Package options')
         .check((args) => {
           if (!args.auto) {
             // single-shot deployment is easy
@@ -151,6 +162,8 @@ module.exports = function deploy() {
         .withFastlyNamespace(argv.fastlyNamespace)
         .withCreatePackages(argv.package)
         .withAddStrain(argv.add)
+        .withMinify(argv.minify)
+        .withResolveGitRefService(argv.svcResolveGitRef)
         .run();
     },